// extensions/openhue/src/api/generated/src/models/GroupedLightPut.ts
export interface GroupedLightPut {
  /**
   * Type of the supported resources (always `grouped_light` here)
   * @type {string}
   * @memberof GroupedLightPut
   */
  type?: GroupedLightPutTypeEnum;
  /**
   *
   * @type {On}
   * @memberof GroupedLightPut
   */
  on?: On;
  /**
   *
   * @type {Dimming}
   * @memberof GroupedLightPut
   */
  dimming?: Dimming;
  /**
   *
   * @type {DimmingDelta}
   * @memberof GroupedLightPut
   */
  dimming_delta?: DimmingDelta;
  /**
   *
   * @type {ColorTemperature}
   * @memberof GroupedLightPut
   */
  color_temperature?: ColorTemperature;
  /**
   *
   * @type {ColorTemperatureDelta}
   * @memberof GroupedLightPut
   */
  color_temperature_delta?: ColorTemperatureDelta;
  /**
   *
   * @type {Color}
   * @memberof GroupedLightPut
   */
  color?: Color;
  /**
   *
   * @type {Alert}
   * @memberof GroupedLightPut
   */
  alert?: Alert;
  /**
   *
   * @type {Signaling}
   * @memberof GroupedLightPut
   */
  signaling?: Signaling;
  /**
   *
   * @type {Dynamics}
   * @memberof GroupedLightPut
   */
  dynamics?: Dynamics;
}

/**
 * @export
 */
export const GroupedLightPutTypeEnum = {
  grouped_light: "grouped_light",
} as const;
export type GroupedLightPutTypeEnum = (typeof GroupedLightPutTypeEnum)[keyof typeof GroupedLightPutTypeEnum];