// extensions/openhue/src/api/generated/src/models/ResourceIdentifier.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hue CLIP API
 * See [openhue/openhue-api](https://github.com/openhue/openhue-api)
 *
 * The version of the OpenAPI document: v2
 * Contact: contact@openhue.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface ResourceIdentifier
 */
export interface ResourceIdentifier {
  /**
   * The unique id of the referenced resource
   * @type {string}
   * @memberof ResourceIdentifier
   */
  rid?: string;
  /**
   * The type of the referenced resource
   * @type {string}
   * @memberof ResourceIdentifier
   */
  rtype?: ResourceIdentifierRtypeEnum;
}

/**
 * @export
 */
export const ResourceIdentifierRtypeEnum = {
  device: "device",
  bridge_home: "bridge_home",
  room: "room",
  zone: "zone",
  light: "light",
  button: "button",
  relative_rotary: "relative_rotary",
  temperature: "temperature",
  light_level: "light_level",
  motion: "motion",
  camera_motion: "camera_motion",
  entertainment: "entertainment",
  contact: "contact",
  tamper: "tamper",
  grouped_light: "grouped_light",
  device_power: "device_power",
  zigbee_bridge_connectivity: "zigbee_bridge_connectivity",
  zigbee_connectivity: "zigbee_connectivity",
  zgp_connectivity: "zgp_connectivity",
  bridge: "bridge",
  zigbee_device_discovery: "zigbee_device_discovery",
  homekit: "homekit",
  matter: "matter",
  matter_fabric: "matter_fabric",
  scene: "scene",
  entertainment_configuration: "entertainment_configuration",
  public_image: "public_image",
  auth_v1: "auth_v1",
  behavior_script: "behavior_script",
  behavior_instance: "behavior_instance",
  geofence: "geofence",
  geofence_client: "geofence_client",
  geolocation: "geolocation",
  smart_scene: "smart_scene",
} as const;
export type ResourceIdentifierRtypeEnum =
  (typeof ResourceIdentifierRtypeEnum)[keyof typeof ResourceIdentifierRtypeEnum];

/**
 * Check if a given object implements the ResourceIdentifier interface.
 */
export function instanceOfResourceIdentifier(value: object): value is ResourceIdentifier {
  return true;
}

export function ResourceIdentifierFromJSON(json: any): ResourceIdentifier {
  return ResourceIdentifierFromJSONTyped(json, false);
}

export function ResourceIdentifierFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResourceIdentifier {
  if (json == null) {
    return json;
  }
  return {
    rid: json["rid"] == null ? undefined : json["rid"],
    rtype: json["rtype"] == null ? undefined : json["rtype"],
  };
}

export function ResourceIdentifierToJSON(json: any): ResourceIdentifier {
  return ResourceIdentifierToJSONTyped(json, false);
}

export function ResourceIdentifierToJSONTyped(
  value?: ResourceIdentifier | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    rid: value["rid"],
    rtype: value["rtype"],
  };
}